Error handling module + php linter

+++ src/Service/TripService.php
@@ -1,9 +1,9 @@
 <?php
 
-
 namespace App\Service;
 
 use App\Entity\Trip;
+use App\Exception\TripNotFoundException;
 use App\Model\TripListItem;
 use App\Model\TripListResponse;
 use App\Repository\TripRepository;
@@ -13,38 +13,41 @@ class TripService
 {
     public function __construct(
         private TripRepository $tripRepository,
-    )
-    {
+    ) {
     }
 
     public function getTripByID(int $id): TripListResponse
     {
         $trip = $this->tripRepository->findOneBy(['id' => $id]);
-        $tripObj = ([(new TripListItem())->setName($trip->getName())
+
+        if (null === $trip) {
+            throw new TripNotFoundException();
+        }
+
+        $tripObj = [(new TripListItem())->setName($trip->getName())
             ->setPrice($trip->getPrice())
             ->setDescription($trip->getDescription())
             ->setDuration($trip->getDuration())
             ->setDateStart($trip->getDateStart()->format('Y-m-d H:i:s'))
-            ->setDateEnd($trip->getDateEnd()->format('Y-m-d H:i:s'))]);
-        return new TripListResponse($tripObj);
+            ->setDateEnd($trip->getDateEnd()->format('Y-m-d H:i:s')), ];
 
+        return new TripListResponse($tripObj);
     }
 
     public function getAllTrips(): TripListResponse
     {
         $trips = $this->tripRepository->findBy([], ['price' => Criteria::ASC]);
         $items = array_map(
-            fn(Trip $trip) => (new TripListItem())
+            fn (Trip $trip) => (new TripListItem())
                 ->setName($trip->getName())
                 ->setDescription($trip->getDescription())
                 ->setDuration($trip->getDuration())
                 ->setPrice($trip->getPrice())
                 ->setDateStart($trip->getDateStart()->format('Y-m-d H:i:s'))
-                ->setDateEnd($trip->getDateEnd()->format('Y-m-d H:i:s'))
-        , $trips);
+                ->setDateEnd($trip->getDateEnd()->format('Y-m-d H:i:s')), $trips);
+
         return new TripListResponse($items);
     }
-
 
     public function save(TripListItem $tripObj): void
     {
@@ -62,7 +65,11 @@ class TripService
     public function delete(int $id): void
     {
         $trip = $this->tripRepository->findOneBy(['id' => $id]);
+
+        if (null === $trip) {
+            throw new TripNotFoundException();
+        }
+
         $this->tripRepository->remove($trip, true);
     }
-
 }
